Revert "Thema 8 Week 2"
