+++ package.json
@@ -14,13 +14,9 @@
     "console": "^0.7.2",
     "core-js": "^2.6.5",
     "csv-parser": "^2.3.1",
-    "d3": "^5.12.0",
-    "d3-vs": "^1.1.1",
     "node-correlation": "^1.0.2",
     "papaparse": "^5.1.0",
-    "sass-loader": "^8.0.0",
     "vue": "^2.6.10",
-    "vue-d3": "^0.1.0",
     "vue-router": "^3.1.3",
     "vue-xls-csv-parser": "0.0.21"
   },
