--- src/Kmeans.vue
@@ -1,26 +1,0 @@
-<template>
-  <div>
-    
-  </div>
-</template>
-
-<script>
-import * as d3 from "d3";
-
-export default {
-    name: "Kmeans",
-    props: [],
-    data() {
-        return {
-            data: {}
-        }
-    },
-    methods: {
-        
-    }
-};
-</script>
-
-<style scoped>
-
-</style>
