+++ src/TextReader.vue
@@ -20,6 +20,6 @@ export default {
 };
 </script>
 
-<style scoped>
+<style>
 
 </style>
